:art: rename take to when
