+++ src/index.js
@@ -1,4 +1,4 @@
 export { applyMiddleware, compose } from 'redux'
 export { default as createStore } from './createStore'
 export * from './factoryHelpers'
-export { default as take } from './take'
+export { default as when } from './when'
